feat: :rocket: Initial implementation of the cache library in TypeScript
Added the initial implementation of the cache library in TypeScript. The library includes basic functionalities such as adding, retrieving, deleting, and clearing cache items, as well as the ability to set time-to-live (TTL) for cache items. Unit tests were provided to ensure the proper functioning of the library. Additionally, instructions were included in the README file on how to install, use, and contribute to the library.